Bump version for first stable release v1.0.0
This contains a breaking change for the target `debian`. Because some
projects use the makefiles to create a debian package without binaries
(f. i. for shell scripts, etc) this target is designed to only create
the actual .deb file.

Instead of calling directly `debian` (which creates a debian file)
use either the phony target `package` or the new target
`debian-with-binary`.

Signed-off-by: Philipp Pixel <pixeloffice@philipppixel.de>