#12 Adhere to debian conflict mgmt for config files; add new lint target
During the upgrade of a Debian package configuration files may change
on both sides: the host, and the new package. This will lead to a
conflict which the package management tool is able to resolve in very
simple terms (f. i. no merge is possible).

The way how the `cesapp` Debian package is built does not support this
way of conflict solution. Instead, the package's configuration files
overwrite any changes made to the config files on the host.

This can lead to CES instances which will not be able to cope with non-
matching configurations up to that the CES instance will cease to
function properly.

This commit changes the build process by adding file that lists the
configuration files so that:

1. the Debian package management can recognize a set of given files as
config files
2. the Debian package management can try to resolve config file conflicts

It is necessary to change the order of build steps in the make target
`$(DEBIAN_PACKAGE)` so that the archives are created when all files
reside in the work directory.

Furthermore this adds a debian package lint target

Signed-off-by: Philipp Pixel <pixeloffice@philipppixel.de>