#28 Support real integration test target; move update target to new file
# Support integration tests

Previously only a mock target existed. This implementation adds integration
testing with the target `integration-test`. This leads to reading a special
list of go packages from the project. This package list is overwritable in order
to support a project-specific scoping.

Due to the nature of the upgrade process there was never a process of getting
rid of old files. That's why a self-update is now implemented.

# Support migration from one repo version to another

The Update routine for updating the framework is moved to its own filial
makefile. The routine works now like this:

1. download makefile repo in target version
2. remove all old makefiles
   - only those which origin from this framework
   - this is due to the nature to work on obsolete filial makefiles which should
error in fail-fast way during the build.
3. move the new files to the original destination