#30 Move deb variable to debian releated makefile
The omnipresent filial makefile `variables.mk` creates a variable for a `deb`
directory. This directory may not be present in every repository.

This commit fixes an error which results from repositories which use the
makefile set but do not use the debian package mechanism. Instead the variable
creation moves to `package-debian.mk`, the only place where it is used.