#34 Optionalize go-junit-report installation; tests depend now on source files
`go-junit-report` is a module that converts golang's test report to the wide-
spread JUnit-compatible report format for further CI/CD usage. If absent, this
module must be installed automatically in order to provide the same test
procedure for developers as well as during CI.

This commit removes a forced installation of the `go-junit-report` module as it
may lead to errors during builds in the Golang build container. That container
has already `go-junit-report` installed, although with root permissions while
the make commands are executed with a non-root user ID. The container would then
not allow the forced installation of the tool again with different permissions.

Additionally we check now whether the sources have changed since the last call
of any test goals. This is done to follow `make`'s idea of only running goals
when these are necessary to be run.

Co-authored-by: Josha von Gizycki <josha.gizycki@cloudogu.com>