#34 check if pre and post steps are empty
if these variables in test-integration and test-unit are empy, make would be executed with an empty parameter, which would result in a call to the default target.
Obviously, this is not what is intended.

Also, GNU make would tolerate this, but IDE support is 100 times better when variables are dereferenced using () instead of {}