Introduce GO_ENV_VARS for go environment variables; #41
The GO_ENV_VARS variable can be set in the projects Makefile
to define go environment variables which should be
set at go build time. This, for example, has been the issue
with the "CGO_ENABLED=0" variable in latest versions of
our Makefiles.

Attention:
The "CGO_ENABLED=0" variable is not set per default any more!
You have to set it via the Makefile, if needed.