#45 create a new variable to control which apt repo is used
Creating a second file which overrides existing targets on include would violate make guidelines and trigger warnings.

An extra target for premium deploy would be error prone as the artifact could still be published to the public repos.

A boolean variable would be the lowest case of declarative style. So, it became an enum-like case of a variable with the possible values `ces-premium` and anything else. It is also future proof!