#48 Add target for cleaning the debian building directory
Building any debian files may leave the debian build directory in any state,
that is: with existing files and directories with any file permissions.

Previously, cleaning the debian build directory was up to user. Yet, the
makefile `package-debian.mk` is responsible to create this state, and
it should also define a respective deb-clean target.

This commit adds a new cleaning target `clean-deb` which removes recursively
the debian building directory.