#68 Extend release shell scripts with extension points
During the release process it might be necessary to modify version within files
that are not tracked by the makefiles (files like Dockerfile, pom.xml, Makefile,
and so on). It also makes sense to not extend these files in all eternity but
rather create true extensibility.

This commit adds this extensibility by having a fixed file in the project's
workspace: `release_args.sh`. This file may define shell functions which are
optionally called (if found) by `release_functions.sh`.