Remove errors by cleaning up actual makefile to run internal shell tests
Originally, the main `Makefile` was intended for grabs. Now that the shell
script complexity surges it is necessary to test these scripts in an automated
way - of course with make itself!

The current state of the Makefile produces lots of errors because exlusive
targets overwrite themselves.

This commit moves the original Makefile as a code snippet into the README.md
and uses only necessary parts to run tests or provide an entry point for
automatic release.