#235 Generate proper xUnit reports with bats
After a breaking change, BATS did no longer generate report files
locally, presumably also not on CI. As BATS introduced another format
switch, this commit does so as well.

The problem is that the container runs as root, so all files generated
have root permissions as well. For a quick remedy, this commit amends
the customBatsEntrypoint.sh file in such a way, that all files in the
BATS report dir are re-owned to user:group 1000:1000.